pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "proj"
version = "0.0.1"
description = "Project for the MLOps Course 2026"
authors = [
  { name = "AndreaZanasi", email = "your@email.com" },
  { name = "DavideMecugni", email = "s253128@dtu.dk"}
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets<3.0",
    "fastapi==0.115.6",
    "hydra-core>=1.3.2",
    "librosa>=0.11.0",
    "lightning>=2.6.0",
    "matplotlib>=3.10.8",
    "pytest>=8.3.4",
    "ruff>=0.1.3",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "torch<2.6.0",
    "torchaudio<2.6.0",
    "torchvision<0.21.0",
    "tqdm>=4.67.1",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb>=0.24.0",
    "click<8.1.0",
    "dvc[gs]>=2.58.2",
    "pathspec==0.11.2"
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = [
    "tests/*",
    "/tmp/*",
    "**/tmp*",
    "**/_remote_module_non_scriptable.py"
]
